refactor(cart,post): kafka listener 에서 DB 접근 기능 분리
